+++ ctng-dbg-compilers-activation-feedstock/recipe/deactivate-gfortran.sh
@@ -83,16 +83,18 @@ function _tc_activation() {
 
 # When people are using conda-build, assume that adding rpath during build, and pointing at
 #    the host env's includes and libs is helpful default behavior
-if [ "${CONDA_BUILD}" = "1" ]; then
+if [ "${CONDA_BUILD:-0}" = "1" ]; then
   FFLAGS_USED="@FFLAGS@ -I${PREFIX}/include -fdebug-prefix-map=\${SRC_DIR}=/usr/local/src/conda/\${PKG_NAME}-\${PKG_VERSION} -fdebug-prefix-map=\${PREFIX}=/usr/local/src/conda-prefix"
 else
   FFLAGS_USED="@FFLAGS@"
 fi
 
-if [ -f /tmp/old-env-$$.txt ]; then
-  rm -f /tmp/old-env-$$.txt || true
+if [ "${CONDA_BUILD:-0}" = "1" ]; then
+  if [ -f /tmp/old-env-$$.txt ]; then
+    rm -f /tmp/old-env-$$.txt || true
+  fi
+  env > /tmp/old-env-$$.txt
 fi
-env > /tmp/old-env-$$.txt
 
 _tc_activation \
   deactivate host @CHOST@ @CHOST@- \
@@ -112,12 +114,14 @@ if [ $? -ne 0 ]; then
   echo "ERROR: $(_get_sourced_filename) failed, see above for details"
 #exit 1
 else
-  if [ -f /tmp/new-env-$$.txt ]; then
-    rm -f /tmp/new-env-$$.txt || true
-  fi
-  env > /tmp/new-env-$$.txt
+  if [ "${CONDA_BUILD:-0}" = "1" ]; then
+    if [ -f /tmp/new-env-$$.txt ]; then
+      rm -f /tmp/new-env-$$.txt || true
+    fi
+    env > /tmp/new-env-$$.txt
 
-  echo "INFO: $(_get_sourced_filename) made the following environmental changes:"
-  diff -U 0 -rN /tmp/old-env-$$.txt /tmp/new-env-$$.txt | tail -n +4 | grep "^-.*\|^+.*" | grep -v "CONDA_BACKUP_" | sort
+    echo "INFO: $(_get_sourced_filename) made the following environmental changes:"
+    diff -U 0 -rN /tmp/old-env-$$.txt /tmp/new-env-$$.txt | tail -n +4 | grep "^-.*\|^+.*" | grep -v "CONDA_BACKUP_" | sort
+    rm -f /tmp/old-env-$$.txt /tmp/new-env-$$.txt || true
+  fi
 fi
-rm -f /tmp/old-env-$$.txt /tmp/new-env-$$.txt || true
